refactor(session): restore backend module budgets
